start move to pyinstaller 6

+++ pyupdater/utils/__init__.py
@@ -290,7 +290,12 @@ def make_archive(name, target, version, archive_format):
         if system.get_system() == 'win':
             src_manifest = src_executable + '.manifest'
             dst_manifest = dst_executable + '.manifest'
-            shutil.move(src_manifest, dst_manifest)
+
+            # check if src manifest exists
+            if os.path.exists(src_manifest):
+                shutil.move(src_manifest, dst_manifest)
+            else:
+                log.warning('Manifest file not found (not sure if this matters): %s', src_manifest)
 
     file_dir = os.path.dirname(os.path.abspath(target))
     filename = '{}-{}-{}'.format(
